feat: new feature for version 0.3.0
- Allowing to receive the transformer input data on an output ensurer
- Allowing to currying functional transformer definition to work like a transformers factory

+++ tests/lib/ensurers.py
@@ -6,7 +6,7 @@ def is_even(num: int):
         raise NumberIsOdd()
 
 
-def is_greater_than_10(num: int):
+def is_greater_than_10(num: float):
     if num <= 10:
         raise NumberLessThanOrEquals10()
